feat: add modelscope download
