+++ pyproject.toml
@@ -20,6 +20,7 @@ dependencies = [
   "msgpack>=1.0.7",
   "safetensors>=0.5.1",
   "huggingface-hub",
+  "modelscope",
   "transformers>=4.57.1",
   "jinja2>=3.1.0",
   "numpy>=1.26",
@@ -69,7 +70,6 @@ benchmark = [
   "tqdm",
   "datasets",
   "pillow",
-  "modelscope",
 ]
 
 dev = [
